pyproject.toml
[tool.poetry]
name = "Jeren Tetris"
version = "0.1.0"
description = ""
authors = ["Jere"]

[tool.poetry.dependencies]
python = "^3.6"
pygame = "^2.1.0"
invoke = "^1.6.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.1.2"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
